requirements.txt
jwt
sqlalchemy
fastapi
bcrypt
django
Pillow
djangorestframework
openai
reportlab
pandas
numpy
scikit-learn
django-cors-headers
python-dotenv
mysqlclient